skills/proposal-supabase-sync/scripts/requirements.txt
pdfplumber>=0.11.0
google-genai>=0.8.0
python-dotenv>=1.0.1
requests>=2.32.0
